Revert "Update to Webpack"
