+++ .eslintrc.json
@@ -12,15 +12,12 @@
 	"settings": {
 		"import/resolver": {
 			"node": {
-				"paths": ["dev/tools", "src"],
+				"paths": ["dev/tools"],
 				"extensions": [".ts"]
 			}
 		}
 	},
     "parser": "@typescript-eslint/parser",
-	"parserOptions": {
-		"warnOnUnsupportedTypeScriptVersion": false
-	},
 	"root": true,
     "rules": {
 		"require-jsdoc": "off",
@@ -35,14 +32,6 @@
 				"external",
 				"type"
 			],
-			"pathGroups": [
-				{
-					// import/order thinks canvas is a builtin module
-					// when only production dependencies are installed
-					"pattern": "canvas",
-					"group": "external"
-				}
-			],
 			"newlines-between": "always",
 			"alphabetize": {
 				"order": "asc",
@@ -50,14 +39,7 @@
 			}
 		}],
 		"max-len": "off",
-		"linebreak-style": "off",
-		"new-cap": "off",
-		"no-restricted-imports": "off",
-		"@typescript-eslint/no-restricted-imports": [
-			"warn",
-			{
-				"patterns": ["../*", "*types"]
-			}
-		]
+		"linebreak-style": ["warn", "windows"],
+		"new-cap": "off"
 	}
 }
