--- .github/actions/initialise-project/action.yml
@@ -1,13 +1,0 @@
-runs:
-  using: composite
-  steps:
-    - name: Setup Node v18.16.0
-      uses: actions/setup-node@v3
-      with:
-        node-version: "18.16.0"
-    - name: Install Yarn
-      shell: bash
-      run: npm install -g yarn
-    - name: Install Dependencies
-      shell: bash
-      run: yarn install --production
