--- .github/actions/lint-project/action.yml
@@ -1,8 +1,0 @@
-runs:
-  using: composite
-  steps:
-    - name: Initialise Project
-      uses: ./.github/actions/initialise-project
-    - name: Lint Project
-      shell: bash
-      run: yarn lint
