+++ .github/workflows/deploy.yml
@@ -1,10 +1,23 @@
-name: Deploy
+name: Build and deploy content to Pages
 
 on:
   push:
-    branches:
-      - main
-      - staging
+    tags: ["v*"]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
+# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
+concurrency:
+  group: "pages"
+  cancel-in-progress: false
 
 jobs:
   build:
@@ -12,75 +25,31 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v3
-      - name: Build Project
-        uses: ./.github/actions/build-project
-
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Build Project
-        uses: ./.github/actions/lint-project
-
-  deploy-staging:
-    runs-on: ubuntu-latest
-    if: github.ref == 'refs/heads/staging'
-    env:
-      VERCEL_ORG_ID: ${{ secrets.VERCEL_ORG_ID }}
-      VERCEL_PROJECT_ID: ${{ secrets.VERCEL_PROJECT_ID }}
-      STAGING_BUILD: true
-    needs: [build, lint]
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Setup Node v18.16.0
-        uses: actions/setup-node@v3
+      - name: Setup Pages
+        uses: actions/configure-pages@v3
+      - name: Setup Go
+        uses: actions/setup-go@v3
         with:
-          node-version: "18.16.0"
-      - name: Install Vercel CLI
-        run: npm install --global vercel@latest
-      - name: Pull Vercel Environment Information
-        run: vercel pull --yes --environment=preview --token=${{ secrets.VERCEL_TOKEN }}
-      - name: Build Project Artifacts
-        run: vercel build --token=${{ secrets.VERCEL_TOKEN }}
-      - name: Deploy Project Artifacts to Vercel
-        run: vercel deploy --prebuilt --token=${{ secrets.VERCEL_TOKEN }}
-
-  create-tag:
-    runs-on: ubuntu-latest
-    if: github.ref == 'refs/heads/main'
-    needs: [build, lint]
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Get Version
-        id: get_version
-        run: echo version="v$(node -p "require('./package.json').version")" >> $GITHUB_OUTPUT
-      - name: Create Tag
-        uses: rickstaa/action-create-tag@v1
+          go-version: 1.19
+      - name: Build Go Project
+        run: |
+          cd ./build
+          go build .
+      - name: Build Public Directory
+        run: |
+          ./build/build -app-version ${{github.ref_name}}
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v1
         with:
-          tag: ${{ steps.get_version.outputs.version }}
+          path: './out'
 
-  deploy-production:
+  deploy:
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
-    if: github.ref == 'refs/heads/main'
-    env:
-      VERCEL_ORG_ID: ${{ secrets.VERCEL_ORG_ID }}
-      VERCEL_PROJECT_ID: ${{ secrets.VERCEL_PROJECT_ID }}
-    needs: [build, lint]
+    needs: build
     steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Setup Node v18.16.0
-        uses: actions/setup-node@v3
-        with:
-          node-version: "18.16.0"
-      - name: Install Vercel CLI
-        run: npm install --global vercel@latest
-      - name: Pull Vercel Environment Information
-        run: vercel pull --yes --environment=production --token=${{ secrets.VERCEL_TOKEN }}
-      - name: Build Project Artifacts
-        run: vercel build --prod --token=${{ secrets.VERCEL_TOKEN }}
-      - name: Deploy Project Artifacts to Vercel
-        run: vercel deploy --prebuilt --prod --token=${{ secrets.VERCEL_TOKEN }}
+      - name: Deploy to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@v2
