--- .github/workflows/test.yml
@@ -1,24 +1,0 @@
-name: Test
-
-on:
-  pull_request:
-    branches:
-      - main
-      - staging
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Build Project
-        uses: ./.github/actions/build-project
-
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Build Project
-        uses: ./.github/actions/lint-project
