+++ .gitignore
@@ -1,3 +1,2 @@
+out
 node_modules
-dist
-cache
