+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+	"liveServer.settings.root": "/public",
+	"liveServer.settings.file": "/404.html"
+}
