+++ TODO.md
@@ -25,32 +25,7 @@
   - [ ] Allow an image to be downloaded of the shape (could link into the image cache)
 - [ ] Display numbers on grid + change frequency of grid lines as you zoom in/out
 - [ ] Allow user to change the colour of the points, depending on the vertex they were generated from
-
-
-## For release [v2.0.0]
-
-### Bugs
-
-- [x] Shape name doesn't have a space next to it in the input label
-  - Fixed by setting all the text in javascript
-- [x] The total time doesn't update in the playback settings
-  - Fixed by making sure the `setInputValue` function fires the event for ranges
-- [x] The incorrect options are displayed when on regular shapes
-  - Fixed by enabling `invert` on the `makeClassToggler` function
-- [x] Inputs don't trigger updates in the graph
-  - Fixed by moving away from old system and declaring which inputs should trigger updates
-- [x] Playback keybinds don't work
-  - Fixed by changing the space code from ` ` to `Space`
-- [x] Time can go above the max time if the delta is very large
-  - Fixed by limiting how big the current time can be
-- [x] The setup state isn't updated when points count or line proportion is changed
-  - Fixed by changing the shape settings input handlers
-- [x] Changing playback speed can cause points to be displayed when they shouldn't
-  - Caused by previous change, fixed by not showing any points when total time is 0
-
-### Features
-
-- [x] Migrate project to typescript and compile with webpack
+- [ ] Migrate project to typescript and compile with webpack
 
 
 ## For release [v1.3.0](https://github.com/Toffee1347/chaos-game/blob/main/CHANGELOG.md#v130)
