+++ dev/tools/run-static-server.ts
@@ -1,4 +1,3 @@
-import {createServer, compileWebpack} from './utils/sandbox.js';
+import {createServer} from './utils/sandbox.js';
 
-compileWebpack();
 createServer();
