--- dev/tools/utils/wait.ts
@@ -1,5 +1,0 @@
-export default function wait(ms: number) {
-	return new Promise((resolve) => {
-		setTimeout(resolve, ms);
-	});
-}
