+++ dev/tsconfig.tools.json
@@ -8,5 +8,8 @@
 	},
 	"include": [
 		"tools/**/*.ts"
+	],
+	"exclude": [
+		"node_modules"
 	]
-}
+}
