+++ package.json
@@ -5,39 +5,20 @@
   "repository": "https://github.com/Toffee1347/chaos-game",
   "author": "Toffee1347 <danielnhoward93@gmail.com>",
   "license": "MIT",
-  "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "copy-webpack-plugin": "^11.0.0",
-    "css-loader": "^6.8.1",
-    "css-minimizer-webpack-plugin": "^5.0.0",
-    "eslint": "^8.42.0",
-    "eslint-config-google": "^0.14.0",
-    "eslint-plugin-import": "^2.27.5",
-    "html-webpack-plugin": "^5.5.1",
-    "mini-css-extract-plugin": "^2.7.6",
-    "ts-loader": "^9.4.3",
-    "typescript": "^5.1.3",
-    "webpack": "^5.86.0",
-    "webpack-cli": "^5.1.4"
-  },
+  "dependencies": {},
   "type": "module",
   "devDependencies": {
     "@types/express": "^4.17.17",
     "@types/gifencoder": "^2.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "@typescript-eslint/parser": "^5.59.9",
     "canvas": "^2.11.2",
+    "eslint": "^8.42.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-import": "^2.27.5",
     "express": "^4.18.2",
     "gifencoder": "^2.0.1",
     "puppeteer": "^20.5.0",
-    "webpack-dev-server": "^4.15.1"
-  },
-  "scripts": {
-    "build": "env NODE_ENV=production webpack build",
-    "dev": "webpack serve",
-    "lint": "eslint --ext ts --max-warnings 0 src dev",
-    "tool": "bash ./dev/run-tool"
-  },
-  "engines": {
-    "node": "18.x"
+    "typescript": "^5.1.3"
   }
 }
