+++ public/404.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<title>404 - Not Found</title>
+</head>
+<body>
+	<script>
+		window.location.replace('/');
+	</script>
+</body>
+</html>
