+++ public/static/css/style.css
@@ -62,26 +62,11 @@ body {
 	position: sticky;
 }
 
+
 .side-overlay-help {
 	transition-delay: 0s;
 	width: 600px;
 	z-index: 200;
-}
-
-@media screen and (min-width: 769px) {
-	.settings-box::-webkit-scrollbar, .type-selection::-webkit-scrollbar {
-		width: 6px;
-		height: 6px;
-	}
-
-	.settings-box::-webkit-scrollbar-thumb, .type-selection::-webkit-scrollbar-thumb {
-		background: rgba(0,0,0,.25);
-		border-radius: 10px;
-	}
-
-	.settings-box::-webkit-scrollbar-thumb:active, .type-selection::-webkit-scrollbar-thumb:active {
-		background: rgba(0,0,0,.4);
-	}
 }
 
 @media screen and (max-width: 768px) {
@@ -138,12 +123,11 @@ body {
 	min-width: max-content;
 	cursor: pointer;
 	box-shadow: 0 6px 10px rgba(0,0,0,.08), 0 0 6px rgba(0,0,0,.05);
-	transition: .3s transform cubic-bezier(.155,1.105,.295,1.12), .3s box-shadow, .3s background-color;
+	transition: .3s transform cubic-bezier(.155,1.105,.295,1.12), .3s box-shadow;
 }
 .type-selection .card:hover {
 	transform: scale(1.05);
 	box-shadow: 0 10px 20px rgba(0,0,0,.12), 0 4px 8px rgba(0,0,0,.06);
-	background-color: #e2e2e2;
 }
 .type-selection .card:last-of-type {
 	margin-right: 0;
