--- tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "ES2022",
-		"target": "ES2022",
-		"lib": [
-			"DOM"
-		]
-	},
-	"include": [
-		"src/**/*.ts"
-	],
-	"exclude": [
-		"node_modules"
-	]
-}
