--- vercel.json
@@ -1,5 +1,0 @@
-{
-	"installCommand": "yarn install --production",
-	"buildCommand": "yarn build",
-	"outputDirectory": "dist"
-}
